fix encoding bug for lcm message byte fields
We cannot assign python strings to lcm message fields that are
declared as bytes and not strings. This worked in python2 but
does not work in python3.  It fails when lcm attempts to encode
the message with an error at the line:

  >>> buf.write(bytearray(self.data[:self.num_bytes]))

In python3 the bytearray constructor requires an encoding to be
specified when converting python strings.